Step the set-like dict-view walks one item at a time, and root their operands across user code (#1633)
* descroperation: root both compare_slot operands across the set conversion

The set / dict-view arm published `a_set` before converting `b`, leaving
`b` unrooted across the first conversion, which mints a set and runs a
user `__hash__` for every element it stores.  A set is
`try_gc_alloc_stable_raw` and does not move, so the existing note about
`a_set` holds, but a `W_DictView` is `try_gc_alloc_nursery_raw` and does,
and this arm requires one operand to be a view.  Publish both operands
before either conversion, in the scope-bound bracket shape the tuple arm
above uses.

No witness of its own: the window needs a collection inside the first
conversion while the other operand is a young view.


* typedef: root the dict-view operands the set ops hold across user code

`dict_view_as_set_op` materialises the left operand into a set, which
hashes every element and so runs Python, while `rhs` stays a native local
no root walker updates.  A `W_DictViewObject` is `try_gc_alloc_nursery_raw`
and moves, and `dict_view_rset_op` reaches this helper with a view as `rhs`
for every reflected operator.  Publish both operands before the
materialisation, read `rhs` back at its use, and pin the minted set: it is
old-gen and never moves, but nothing else holds it while the named in-place
method runs.

`dict_view_isdisjoint` pinned its receiver only after
`crate::builtins::collect_iterable`, which runs a user `__iter__` and
`__next__`, and the receiver is a view on every path that reaches it.  The
publish moves above that call; the pins on the items it hands back stay
where they were.

No witness of its own: both windows need a collection inside the call while
the operand is a young view.  A `d.keys()` `-` `&` `|` `^` stress with an
allocating user `__hash__` under `PYPY_GC_NURSERY=64K` agrees with cpython
on the pre-fix binary, with and without the JIT.


* Port JitCode lowerer parity fixes

* typedef, descroperation: step the set-like dict-view walks one item at a time

`dictmultiobject.py descr_isdisjoint` opens `space.iter(w_other)` and tests
`space.contains_w(self, w_item)` per step.  `dict_view_isdisjoint` called
`collect_iterable` first, so an operand that yielded a common element and
then raised propagated that exception instead of answering `False`, and a
generator was drained past the answer.  It also had no branch for iterating
the shorter operand when both are set-like.

`dictmultiobject.py _all_contained_in` walks `space.iter(w_dictview)` for the
same reason on the other side.  `dict_view_all_contained_in` read a
`dict_view_snapshot` taken before the loop, so a `contains_w` that changed the
dict's size never reached the view iterator's own size check and the
comparison returned an answer rather than raising.

`compare_slot`'s set-vs-view arm reduced both operands to sets built from that
same snapshot.  It now hands the pair to `dict_view_compare` with the operator
reflected when the view is the right-hand operand, which is the method a set's
`NotImplemented` sends the comparison to upstream.

Both walks hold their operands, the iterator and the yielded item in
shadow-stack slots and read them back at each use; the item slot is rewritten
per turn rather than pinned per item, and a dict-view iterator's source dict is
retained for the native loop as `collect_iterator` retains it.

`dict_view_setlike_lazy_walk_regression.py` is the witness: 8 of its assertions
fail on the pre-fix binary, both before and after the JIT compiles `warm`, and
it passes on CPython 3.14, PyPy and the fixed binary.  `pyre/check.py --backend
dynasm` reports 532/532.


* typedef: root the dict-view comparison operands across the length dispatch

`dict_view_compare` read both operands as native locals across two
`baseobjspace::len` calls and then used them for `dict_view_all_contained_in`.
`len` reaches a user `__len__` whenever an operand is a set subclass that
overrides it, which a witness confirms on every arm:

    class S(set):
        def __len__(self):
            gc.collect(0)
            return set.__len__(self)

    S({1}) == d.keys()      # the override runs
    d.keys() == S({1})      # so does this one

A dict view is `try_gc_alloc_nursery_raw` and so moves, so the second length
call relocated the view the first arm still had to read, and the boxed int the
first call returned was held across the second.  Both operands are now
published before either length call and read back from their slots, and each
length is reduced to its native value where it is taken.

This produces a wrong answer, not just an unsound window.  Two binaries built
from this tree differing only in the hunk below, run interleaved under
`PYPY_GC_NURSERY=64K` over 20000 comparisons of a collecting `__len__` against
a dict view nothing else holds:

    round 1  unrooted: wrong answers: 1   rooted: wrong answers: 0
    round 2  unrooted: wrong answers: 1   rooted: wrong answers: 0
    round 3  unrooted: wrong answers: 1   rooted: wrong answers: 0
    round 4  unrooted: wrong answers: 1   rooted: wrong answers: 0
    round 5  unrooted: wrong answers: 1   rooted: wrong answers: 0

The default nursery reports 0 on both, so the shape is only reachable with the
nursery small enough for the churn to force the relocation — which is why the
fixture below documents the shapes rather than gating on the count.

Reported by the Codex review of 22262b7eb1a, which named the set-subclass
`__len__` window that `descroperation`'s reflected arm newly routes through.

The fixture gains the two shapes the walks did not cover: a set subclass whose
`__len__` collects, on both sides of all six operators, and an iterable whose
`__iter__` collects before returning its iterator -- a generator's `iter()`
runs no user code, so nothing previously exercised the receiver root
`dict_view_isdisjoint` takes before `space.iter(w_other)`.

`pyre/check.py --backend dynasm` reports 532/532.

+++ majit/majit-metainterp/src/jitcode/assembler.rs
@@ -5862,9 +5862,13 @@ impl JitCodeBuilder {
             // descriptor banks are empty.
             str_consts: Vec::new(),
             unit_variant_consts: Vec::new(),
-            c_num_regs_i: self.num_regs_i,
-            c_num_regs_r: self.num_regs_r,
-            c_num_regs_f: self.num_regs_f,
+            // `jitcode.py JitCode.setup`: the three register counts are
+            // stored as one `chr` apiece.  The gate above proves these casts
+            // lossless; keeping the builder counters wide is useful only
+            // while detecting an unencodable source body.
+            c_num_regs_i: self.num_regs_i as u8,
+            c_num_regs_r: self.num_regs_r as u8,
+            c_num_regs_f: self.num_regs_f as u8,
             // RPython `assembler.py make_jitcode(startpoints=
             // self.startpoints, alllabels=self.alllabels, ...)` —
             // assembled jitcodes always carry the recorded set, never
